Add some base container related codes

+++ build.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
 REPOURL="ghcr.io/fjordos/simple"
+FVER=44
 FVERLATEST=44
+FARCH="x86_64"
 
 if [[ -f podman-login.inc ]] || podman login --get-login $REPOURL ; then
   podman login $(cat podman-login.inc) $REPOURL
@@ -30,36 +32,48 @@ if false ; then
   done
 fi
 
+for COMPONENT in $(find . -maxdepth 1 -name "Containerfile.*" | awk -F . '{print $3}') ; do
+  podman image rm -f "$COMPONENT":new-build
+  podman build --logfile "$COMPONENT.log" --tag "$COMPONENT":new-build -f Containerfile."$COMPONENT" .
+  for TAG in ${FVER} ${FVER}-${FARCH} ; do
+    TAGS="$TAGS $REPOURL/$COMPONENT:$TAG $REPOURL/$COMPONENT:$TAG-$(date +%Y-%m-%d)"
+  done
+  TAGS="$TAGS $REPOURL/$COMPONENT:latest"
+  for TAG in $TAGS ; do
+    echo podman tag "$COMPONENT":new-build $TAG
+    podman tag "$COMPONENT":new-build $TAG
+    echo podman push $TAG
+    podman push --sign-by release@fjordos.no $TAG
+  done
+  podman image rm -f "$COMPONENT":new-build
+done
 #exit 0
-for COMPONENT in fjordos-base-minimal $(find . -mindepth 1 -maxdepth 1 -type d -not -name ".*" -not -name "fjordos-*" | sed 's#./##') ; do
-  if [[ ! -f ./$COMPONENT/dont-build ]] ; then
+
+for COMPONENT in $(find . -mindepth 1 -maxdepth 1 -type d -not -name ".*" | sed 's#./##') ; do
+  if [[ ! -f ./$COMPONENT/dont-build && -e ./$COMPONENT/Containerfile ]] ; then
     [[ -e ./$COMPONENT/Dockerfile ]] || ln -s Containerfile ./$COMPONENT/Dockerfile
-    podman build ./$COMPONENT | tee $COMPONENT.log
-    IMG=$(tail -n 1 $COMPONENT.log)
-    if [[ "$COMPONENT" == fjordos-* ]] ; then
-      case "$COMPONENT" in
-      "fjordos-base-minimal")
-        VERS="${FVER} ${FVER}-${FARCH}"
-        ;;
-      "*")
-        VERS="invalid"
-      esac
-    elif [[ "$COMPONENT" == "go" ]] ; then
-      VERS=$()
-    else
+    podman image rm -f "$COMPONENT":new-build
+    podman build --logfile "$COMPONENT.log" --tag "$COMPONENT":new-build "./$COMPONENT"
+    case $COMPONENT in
+    "go")
+      VERS="invalid"
+      ;;
+    "*")
       VERS=$(podman run --rm -ti $IMG rpm -q $COMPONENT | sed "s#$COMPONENT-##" | tr -d "\r")
-    fi
+    ;;
+    esac
     if [[ ${VERS} != "invalid" ]] ; then
         TAGS="$REPOURL/$COMPONENT:latest"
         for TAG in $VERS ; do
           TAGS="$TAGS $REPOURL/$COMPONENT:$TAG $REPOURL/$COMPONENT:$TAG-$(date +%Y-%m-%d)"
         done
       for TAG in $TAGS ; do
-        echo podman tag $IMG $TAG
-        podman tag $IMG $TAG
+        echo podman tag "$COMPONENT":new-build $TAG
+        podman tag "$COMPONENT":new-build $TAG
         echo podman push $TAG
         podman push $TAG
       done
     fi
   fi
+  podman image rm -f "$COMPONENT":new-build
 done
